refactor: Move `is_close` to `CompliantColumn`
Now that #2962 has merged, this part of the plan is possible (https://github.com/narwhals-dev/narwhals/pull/2962#issuecomment-3171900789)
